+++ tests/testthat/_snaps/rix/one_git_default.nix
@@ -30,17 +30,19 @@ let
       nix;
   };
   
-in
-
-pkgs.mkShell {
-  LOCALE_ARCHIVE = if pkgs.system == "x86_64-linux" then "${pkgs.glibcLocales}/lib/locale/locale-archive" else "";
-  LANG = "en_US.UTF-8";
+  shell = pkgs.mkShell {
+    LOCALE_ARCHIVE = if pkgs.system == "x86_64-linux" then "${pkgs.glibcLocales}/lib/locale/locale-archive" else "";
+    LANG = "en_US.UTF-8";
    LC_ALL = "en_US.UTF-8";
    LC_TIME = "en_US.UTF-8";
    LC_MONETARY = "en_US.UTF-8";
    LC_PAPER = "en_US.UTF-8";
    LC_MEASUREMENT = "en_US.UTF-8";
 
-  buildInputs = [ housing   system_packages   ];
-  
-}
+    buildInputs = [ housing   system_packages   ];
+    
+  }; 
+in
+  {
+    inherit pkgs shell;
+  }